Fork move-to-bugzilla as the basis for site-triage addon
This addon is designed to help with the triage of site issues.

The first part is a specialized version of move-to-bugzilla for the
purpose of coverting webcompat/web-bugs GitHub issues into bugzilla
reports in the Web Compatibility :: Site Reports component, using
additional labels and metadata on the GitHub issue to fill in as much
of the bugzilla data automatically as possible.