Fix serialization of default value controls.
The previous change regressed serialization of controls with default
values, because we started computing the default state from the
control itself, even when it wasn't explictly set. Then because we
weren't serializing any controls in their default state we started not
serializing any select controls unless they'd been changed.

With this change you have to explictly provide
`data-serialization-options="no-default"` on controls that shouldn't
serialize their default state.

However getthing the `data-` attribute wasn't really supported for all
controls, since `datasetValue` would look at the current
state (i.e. the selected option, or whether a checkbox is
selected). That's used for serializing the correct keywords. To fix
that rename the existing method to currentDatasetValue and add
datasetValue that doesn't look at the state of the control.

This is still all rather over-complex for something quite simple.