Switch to purely score based priority
In this model anything with a score > 100 is P2 and > 750 is P1. So
for example a dsktop-only broken workflow (80 points) on a local top 1000
site (x10) would be P1, whereas a significant visual issue (20 points)
affecting a local top 10,000 site (x5 multiplier) would be an example
of the minimum P2.